Reorganize ops and run full test suite on build
The main impetus for this is the fact that tests are currently broken on
main. This is because SDK tests, webapp tests, and api validation are
not run during the main build that PR checks use.

Additionally, this consolidates a few ops to the top level opspec. Now
that there is a single go project, this makes more sense to me.