Add additional vscode debugging features + native toolchain instructions
The first new vscode launch configuration allows attaching to a built cli
launched directly from the host, instead of through the dev op.

The second allows running a given test file with an attached debugger.
This lets you avoid modifying the test op to quickly debug a given test
file. Since this project uses Ginko for testing, you don't have the
option to use the vscode go plugin's auto-test detection, so this is a
decent alternative.