Detect invalid outputs in opspecs
This introduces validation of outputs when making op calls. Currently,
if an invalid output is used, it's ignored. When used later in an op,
it's easy to misinterpret the errors produced as happening in the wrong
place.

I made an attempt to try to detect common user-errors and proactively
provide information to help solve the error (see test cases).

This is technically backwards incompatible - currently if an invalid op
output is specified and never used, it will be silently ignored. This
changes so it will now cause an error, as it's not a valid reference.
Resolving is as simple as removing the bad output line, so I think this
is definitely worth the increased usability, given the v0 status of
opctl.

Resolves https://github.com/opctl/opctl/issues/798